making ids deterministic
