--- core/domain-snapshots/outputs/35f3c3c65bc3eb48d30c92755bbf07f4cc1e0548.json
@@ -1,1 +1,0 @@
-[{"ShapeAdded":{"shapeId":"basic_0","baseShapeId":"$object","parameters":{"DynamicParameterList":{"shapeParameterIds":[]}},"name":"","eventContext":{"clientId":"ccc","clientSessionId":"sss","clientCommandBatchId":"bbb","createdAt":"NOW"}}},{"ShapeAdded":{"shapeId":"basic_2","baseShapeId":"$string","parameters":{"DynamicParameterList":{"shapeParameterIds":[]}},"name":"","eventContext":{"clientId":"ccc","clientSessionId":"sss","clientCommandBatchId":"bbb","createdAt":"NOW"}}},{"FieldAdded":{"fieldId":"basic_1","shapeId":"basic_0","name":"a","shapeDescriptor":{"FieldShapeFromShape":{"fieldId":"basic_1","shapeId":"basic_2"}},"eventContext":{"clientId":"ccc","clientSessionId":"sss","clientCommandBatchId":"bbb","createdAt":"NOW"}}},{"ShapeAdded":{"shapeId":"basic_4","baseShapeId":"$string","parameters":{"DynamicParameterList":{"shapeParameterIds":[]}},"name":"","eventContext":{"clientId":"ccc","clientSessionId":"sss","clientCommandBatchId":"bbb","createdAt":"NOW"}}},{"FieldAdded":{"fieldId":"basic_3","shapeId":"basic_0","name":"b","shapeDescriptor":{"FieldShapeFromShape":{"fieldId":"basic_3","shapeId":"basic_4"}},"eventContext":{"clientId":"ccc","clientSessionId":"sss","clientCommandBatchId":"bbb","createdAt":"NOW"}}},{"ShapeAdded":{"shapeId":"basic_6","baseShapeId":"$string","parameters":{"DynamicParameterList":{"shapeParameterIds":[]}},"name":"","eventContext":{"clientId":"ccc","clientSessionId":"sss","clientCommandBatchId":"bbb","createdAt":"NOW"}}},{"FieldAdded":{"fieldId":"basic_5","shapeId":"basic_0","name":"c","shapeDescriptor":{"FieldShapeFromShape":{"fieldId":"basic_5","shapeId":"basic_6"}},"eventContext":{"clientId":"ccc","clientSessionId":"sss","clientCommandBatchId":"bbb","createdAt":"NOW"}}},{"ShapeRenamed":{"shapeId":"basic_0","name":"Basic","eventContext":{"clientId":"ccc","clientSessionId":"sss","clientCommandBatchId":"bbb","createdAt":"NOW"}}}]
