+++ core/optic/shared/src/test/scala/com/useoptic/end_to_end/snapshot_task/CommandsToEvents.scala
@@ -1,0 +1,5 @@
+package com.useoptic.end_to_end.snapshot_task
+
+class CommandsToEvents {
+
+}
